+++ src/Workspaces/SharedUtilitiesAndExtensions/Workspace/Core/CodeFixes/AbstractDocumentBasedFixAllProvider.cs
@@ -20,81 +20,16 @@ namespace Microsoft.CodeAnalysis.CodeFixes
     /// <summary>
     /// Provides a base class to write a <see cref="FixAllProvider"/> that fixes documents independently.
     /// </summary>
-    internal abstract class AbstractDocumentBasedFixAllProvider : FixAllProvider
+    internal abstract class AbstractDocumentBasedFixAllProvider : AbstractDefaultFixAllProvider
     {
-        protected abstract string GetFixAllTitle(FixAllContext fixAllContext);
-
         protected abstract Task<Document?> FixAllAsync(FixAllContext context, Document document, ImmutableArray<Diagnostic> diagnostics);
-
-        public sealed override IEnumerable<FixAllScope> GetSupportedFixAllScopes()
-            => base.GetSupportedFixAllScopes();
-
-        public sealed override async Task<CodeAction?> GetFixAsync(FixAllContext fixAllContext)
-        {
-            Contract.ThrowIfFalse(fixAllContext.Scope is FixAllScope.Document or FixAllScope.Project or FixAllScope.Solution);
-
-            var solution = fixAllContext.Scope switch
-            {
-                FixAllScope.Document => await GetDocumentFixesAsync(fixAllContext).ConfigureAwait(false),
-                FixAllScope.Project => await GetProjectFixesAsync(fixAllContext).ConfigureAwait(false),
-                FixAllScope.Solution => await GetSolutionFixesAsync(fixAllContext).ConfigureAwait(false),
-                _ => throw ExceptionUtilities.UnexpectedValue(fixAllContext.Scope),
-            };
-
-            if (solution == null)
-                return null;
-
-#pragma warning disable RS0005 // Do not use generic 'CodeAction.Create' to create 'CodeAction'
-
-            return CodeAction.Create(
-                GetFixAllTitle(fixAllContext),
-                c => Task.FromResult(solution));
-
-#pragma warning disable RS0005 // Do not use generic 'CodeAction.Create' to create 'CodeAction'
-        }
-
-        private Task<Solution> GetDocumentFixesAsync(FixAllContext fixAllContext)
-            => FixAllContextsAsync(
-                fixAllContext,
-                ImmutableArray.Create(fixAllContext));
-
-        private Task<Solution> GetProjectFixesAsync(FixAllContext fixAllContext)
-            => FixAllContextsAsync(
-                fixAllContext,
-                ImmutableArray.Create(fixAllContext.WithDocument(null)));
-
-        private Task<Solution> GetSolutionFixesAsync(FixAllContext fixAllContext)
-        {
-            var solution = fixAllContext.Solution;
-            var dependencyGraph = solution.GetProjectDependencyGraph();
-
-            // Walk through each project in topological order, determining and applying the diagnostics for each
-            // project.  We do this in topological order so that the compilations for successive projects are readily
-            // available as we just computed them for dependent projects.  If we were to do it out of order, we might
-            // start with a project that has a ton of dependencies, and we'd spend an inordinate amount of time just
-            // building the compilations for it before we could proceed.
-            //
-            // By processing one project at a time, we can also let go of a project once done with it, allowing us to
-            // reclaim lots of the memory so we don't overload the system while processing a large solution.
-            //
-            // Note: we have to filter down to projects of the same language as the FixAllContext points at a
-            // CodeFixProvider, and we can't call into providers of different languages with diagnostics from a
-            // different language.
-            var sortedProjects = dependencyGraph.GetTopologicallySortedProjects()
-                                                .Select(id => solution.GetRequiredProject(id))
-                                                .Where(p => p.Language == fixAllContext.Project.Language);
-            return FixAllContextsAsync(
-                fixAllContext,
-                sortedProjects.SelectAsArray(
-                    p => fixAllContext.WithScope(FixAllScope.Project).WithProject(p).WithDocument(null)));
-        }
 
         /// <summary>
         /// All fix-alls funnel into this method.  For doc-fix-all or project-fix-all call into this with just their
         /// single <see cref="FixAllContext"/> in <paramref name="fixAllContexts"/>.  For solution-fix-all, <paramref
         /// name="fixAllContexts"/> will contain a context for each project in the solution.
         /// </summary>
-        private async Task<Solution> FixAllContextsAsync(
+        protected override sealed async Task<Solution?> FixAllContextsAsync(
             FixAllContext originalFixAllContext,
             ImmutableArray<FixAllContext> fixAllContexts)
         {
